+++ interop/src/facade64.rs
@@ -3,6 +3,7 @@
 //! Please refer to the other chapters of the help for documentation of the functions.
 use super::*;
 use basic_dsp_vector::*;
+use basic_dsp_vector::meta;
 use basic_dsp_vector::conv_types::*;
 use num_complex::*;
 use std::slice;
@@ -148,7 +149,7 @@ pub extern "C" fn real_dot_product64(vector: &VecBuf,
                                      operand: &VecBuf)
                                      -> ScalarInteropResult<f64> {
     vector.convert_scalar(|v| {
-        DotProductOps::<GenDspVec<Vec<f64>, f64>, f64, f64, RealOrComplexData, TimeOrFrequencyData>::dot_product(v, &operand.vec)
+        DotProductOps::<GenDspVec<Vec<f64>, f64>, f64, f64, meta::RealOrComplex, meta::TimeOrFreq>::dot_product(v, &operand.vec)
     },
     0.0)
 }
@@ -158,7 +159,7 @@ pub extern "C" fn complex_dot_product64(vector: &VecBuf,
                                         operand: &VecBuf)
                                         -> ScalarInteropResult<Complex64> {
     vector.convert_scalar(|v| {
-        DotProductOps::<GenDspVec<Vec<f64>, f64>, Complex64, f64, RealOrComplexData, TimeOrFrequencyData>::dot_product(v, &operand.vec)
+        DotProductOps::<GenDspVec<Vec<f64>, f64>, Complex64, f64, meta::RealOrComplex, meta::TimeOrFreq>::dot_product(v, &operand.vec)
     },
     Complex64::new(0.0, 0.0))
 }
@@ -199,7 +200,7 @@ pub extern "C" fn real_dot_product_prec64(vector: &VecBuf,
                                           operand: &VecBuf)
                                           -> ScalarInteropResult<f64> {
     vector.convert_scalar(|v| {
-        PreciseDotProductOps::<GenDspVec<Vec<f64>, f64>, f64, f64, RealOrComplexData, TimeOrFrequencyData>::dot_product_prec(v, &operand.vec)
+        PreciseDotProductOps::<GenDspVec<Vec<f64>, f64>, f64, f64, meta::RealOrComplex, meta::TimeOrFreq>::dot_product_prec(v, &operand.vec)
     },
     0.0)
 }
@@ -209,7 +210,7 @@ pub extern "C" fn complex_dot_product_prec64(vector: &VecBuf,
                                              operand: &VecBuf)
                                              -> ScalarInteropResult<Complex64> {
     vector.convert_scalar(|v| {
-        PreciseDotProductOps::<GenDspVec<Vec<f64>, f64>, Complex64, f64, RealOrComplexData, TimeOrFrequencyData>::dot_product_prec(v, &operand.vec)
+        PreciseDotProductOps::<GenDspVec<Vec<f64>, f64>, Complex64, f64, meta::RealOrComplex, meta::TimeOrFreq>::dot_product_prec(v, &operand.vec)
     },
     Complex64::new(0.0, 0.0))
 }